% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam2fragCounts.R
\name{Bam2FragCounts}
\alias{Bam2FragCounts}
\title{Import bam files and generate read counts per GATC fragment}
\usage{
Bam2FragCounts(meta.data, bam.dir, GATC.fragments, maxgap = 0, shift = 0,
  VERBOSE = FALSE, addToName = "")
}
\arguments{
\item{meta.data}{Data frame.}

\item{bam.dir}{Character, directory which contains the bamfiles.}

\item{GATC.fragments}{GRanges object with GATC regions.}

\item{maxgap}{Integer, maximum allowed gap between read alignment position
and GATC position.}

\item{shift}{Integer, exact number of bases by which reads are shifted
before comparing to GATC position.}

\item{VERBOSE}{Logical, print diagnostic msgs [default = FALSE].}

\item{addToName}{Character, string which is added to the chromosome names
read from the bamfile}
}
\value{
GRanges object; granges specify genomic GATC fragments, mcols
specify readcounts (forw/rev/both) per sample
}
\description{
\code{BamToFragCounts} generates read counts per GATC fragment for a set of bam files.
}
\details{
extended description of function
}
