% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEfromBAM.R
\name{bamToOE}
\alias{bamToOE}
\title{Compute OE scores from (a set of) bamfiles.}
\usage{
bamToOE(exp.name, bam.dir = getwd(), bam.fname, sample.name, cell.count,
  sig = sprintf("LP\%s", format(Sys.time(), "\%y\%m\%d")), CHR,
  bin.size = "chromlength", fExp, GATC.mpbl.reads, OE.norm.factor = NULL,
  outdir, VERBOSE = FALSE, do.norm = TRUE)
}
\arguments{
\item{exp.name}{Character, used to create output file names.}

\item{bam.dir}{Character, directory which contains the bamfiles.}

\item{bam.fname}{Character vector, bamfile names for all samples.}

\item{sample.name}{Character vector, (short) names for all samples.}

\item{cell.count}{Integer vector, cell count for all samples.}

\item{sig}{Character, string with initials and date used for output file names (eg LP170619).}

\item{CHR}{Character vector, chromosome names to be used in the analysis.}

\item{bin.size}{integer, 0 (each bin is entire chr
length) or >0 specifying the width in bp of the bins used for computing the
OE scores.}

\item{fExp}{Numeric vector, length of twice the number of GATC fragments in
\code{GATC.mpbl.reads} (2 reads per fragment) specifying the (normalized)
expected readcount per fragment.}

\item{GATC.mpbl.reads}{GRanges object, granges specify genomic GATC fragments, mcols specifies whether forw/rev/both read count (0/1/2).}

\item{OE.norm.factor}{Numeric vector, normalization factor per window/chromosome to normalize raw OE scores.}

\item{outdir}{Character, specifies directory for output files.}

\item{VERBOSE}{Logical, print diagnostic msgs [default = FALSE].}

\item{do.norm}{Logical, perform normalization [default = FALSE].}
}
\value{
List with 3 elements;
\describe{
  \item{OE.norm.factor}{Numeric vector; normalization factor per window/chromosome to normalize raw OE scores}
  \item{OE.norm}{GRanges object; granges specify genomic GATC fragments, mcols specify normalized OE scores for each sample (ie each bamfile)}
  \item{OE.raw}{GRanges object; granges specify genomic GATC fragments, mcols specify raw OE scores for each sample (ie each bamfile)}
}
}
\description{
\code{bamToOE} computes the observed-over-expected (OE) scores for a set of samples. The input is a bamfile for each sample.
}
\details{
extended description of function
}
\examples{
\dontrun{
res <- bamToOE(exp.name="myExperiment", bam.dir="./bamfiles/",
  bam.fname=c("1.bam","2.bam"), sample.name=c("smpl1","smpl2"),
  cell.count=c(20,1),CHR=paste0('chr', c(1:19, 'X')), fExp=fExp.mm10,
  GATC.mpbl.reads=GATC.mpbl.reads.mm10, outdir="myResults")
}

}
