% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fExp.mm10.R
\docType{data}
\name{fExp.mm10}
\alias{fExp.mm10}
\title{Expected readcounts (forw/rev) for mm10 GATC fragments}
\format{A numric vector with length 2x number of GATC fragments in mm10}
\usage{
fExp.mm10
}
\description{
Expected readcounts (forw/rev) for mm10 GATC fragments
}
\keyword{datasets}
