% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GATC.mpbl.reads.mm10.R
\docType{data}
\name{GATC.mpbl.reads.mm10}
\alias{GATC.mpbl.reads.mm10}
\title{GRanges object with all GATC fragments of mm10, plus mappable reads.}
\format{A GRanges object; \code{granges} specify GATC fragments.
\code{mcols} has three integer columns with mappable read counts for
forward/reverse/both strands.}
\usage{
GATC.mpbl.reads.mm10
}
\description{
The reads are 39 bp
}
\keyword{datasets}
