% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fragLen_seqCompos_correction.R
\name{PreprocReadcounts}
\alias{PreprocReadcounts}
\title{Preprocess readcounts in GATC.readcounts}
\usage{
PreprocReadcounts(GATC.readcounts)
}
\arguments{
\item{GATC.readcounts}{GRanges object with GATC regions and readcounts (forw/rev/both) per sample.}
}
\value{
GRanges object; granges specify genomic GATC fragments, mcols specify readcounts (forw/rev/both) per sample
}
\description{
\code{PreprocReadcounts} preprocesses readcounts, ie flatten counts to 0/1, etc
}
\details{
Details: extended description of function
}

