% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fragLen_seqCompos_correction.R
\name{SummarizeReadcounts}
\alias{SummarizeReadcounts}
\title{Compute OE scores from read counts}
\usage{
SummarizeReadcounts(fExp, GATC.readcounts, binsize)
}
\arguments{
\item{fExp}{Numeric vector, length 2x number of GATC fragments, specifies
for each possible forward and reverse read whether it is expected (ie,
mappable).}

\item{GATC.readcounts}{GRanges object with GATC regions and readcounts (forw/rev/both) per sample.}

\item{binsize}{Integer, either a single value or a vector of same length as number of chromosomes.}
}
\value{
GRanges object; granges specify genomic regions of size \code{bin.size}, mcols specify computed OE scores
}
\description{
\code{SummarizeReadcounts} computes OE scores from read counts
}
\details{
Details: extended description of function
}

